Code refactoring. Moving letter picking logic into its own file
